backend/requirements.txt
# Python Backend Dependencies
# Install: pip install -r requirements.txt

# FastAPI & Server
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Supabase Client
supabase==2.28.0

# LLM & AI
google-generativeai==0.4.0
groq==0.4.2
pillow==10.2.0
pytesseract==0.3.10

# Data Validation
pydantic>=2.6.1
pydantic-settings==2.1.0

# API Requests
httpx>=0.26,<0.27
requests==2.32.3

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Date handling
python-dateutil==2.8.2

# JSON Schema validation
jsonschema==4.21.1

# Image processing (for receipt OCR)
pytesseract==0.3.10

# Caching
cachetools>=5.3.2

# CORS
fastapi-cors==0.0.6
pyasn1<0.5.0,>=0.4.6